+++ src/pages/PremiacaoReconquista/PremiacaoReconquista.js
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from 'react';
-import { Box, Typography, Button, Paper, CircularProgress, Alert, TextField, Grid, Dialog, DialogActions, DialogContent, DialogTitle } from '@mui/material';
+import { Box, Typography, Button, Paper, CircularProgress, Alert, TextField, Grid, Dialog, MenuItem, DialogActions, DialogContent, DialogTitle, DialogContentText } from '@mui/material';
 import { FaPencilAlt, FaTrashAlt } from "react-icons/fa";
 import { DataGrid } from '@mui/x-data-grid';
 import SidebarMenu from '../../components/SidebarMenu';
@@ -8,6 +8,7 @@ import { useNavigate } from 'react-router-dom';
 import { getFilteredPremiacaoReconquistaData, deletePremiacaoReconquista, updatePremiacaoReconquista } from '../../services/apiService';
 import jsPDF from 'jspdf';
 import autoTable from 'jspdf-autotable';
+import { useToast } from '../../components/ToastProvider';
 
 const PremiacaoReconquista = ({ toggleTheme }) => {
   const [sidebarOpen, setSidebarOpen] = useState(true);
@@ -18,12 +19,14 @@ const PremiacaoReconquista = ({ toggleTheme }) => {
   const [filterDescricao, setFilterDescricao] = useState('');
   const [filterTime, setFilterTime] = useState('');
   const [filterValor, setFilterValor] = useState('');
-
+  const { showToast } = useToast();
   const [successMessage, setSuccessMessage] = useState('');
   const [errorMessage, setErrorMessage] = useState('');
   const [editingPremiacaoReconquista, setEditingPremiacaoReconquista] = useState(null);
   const [openModal, setOpenModal] = useState(false);
-
+  const [deleteDialogOpen, setDeleteDialogOpen] = useState(false);
+  const [itemToDelete, setItemToDelete] = useState(null);
+  const [openConfirmDialog, setOpenConfirmDialog] = useState(false);
   const navigate = useNavigate();
 
   useEffect(() => {
@@ -34,8 +37,8 @@ const PremiacaoReconquista = ({ toggleTheme }) => {
         setData(result);
         setFilteredData(result);
       } catch (error) {
-        setError('Erro ao buscar dados de premiacaoReconquista.');
-        console.error('Erro ao buscar dados de premiacaoReconquista:', error);
+        showToast('Erro ao buscar dados de Premiação Reconquista.', 'error');
+        console.error('Erro ao buscar dados de Premiação Reconquista:', error);
       } finally {
         setLoading(false);
       }
@@ -48,7 +51,7 @@ const PremiacaoReconquista = ({ toggleTheme }) => {
   }, [filterDescricao, filterTime, filterValor, data]);
 
   const applyFilters = () => {
-    let filtered = [...data]; 
+    let filtered = [...data];
     if (filterDescricao) {
       filtered = filtered.filter((premiacaoReconquista) =>
         premiacaoReconquista.descricao.toLowerCase().includes(filterDescricao.toLowerCase())
@@ -86,9 +89,22 @@ const PremiacaoReconquista = ({ toggleTheme }) => {
     setOpenModal(false);
     setEditingPremiacaoReconquista(null);
   };
-
-  const handleSaveClick = async () => {
+  const handleSaveClick = () => {
+    setOpenConfirmDialog(true); // Abrir modal de confirmação
+  };
+  const handleConfirmSave = async () => {
     if (editingPremiacaoReconquista) {
+      // Validation logic
+      if (!editingPremiacaoReconquista.descricao ||
+        !editingPremiacaoReconquista.time ||
+        editingPremiacaoReconquista.valor < 0 ||
+        editingPremiacaoReconquista.minimo < 0 ||
+        editingPremiacaoReconquista.maximo < 0) {
+        showToast('Por favor, preencha todos os campos corretamente.', 'error');
+
+        return; // Exit the function if validation fails
+      }
+
       try {
         const updatedData = {
           descricao: editingPremiacaoReconquista.descricao,
@@ -99,7 +115,9 @@ const PremiacaoReconquista = ({ toggleTheme }) => {
         };
 
         await updatePremiacaoReconquista(editingPremiacaoReconquista.descricao, updatedData);
-        setSuccessMessage('PremiacaoReconquista atualizado com sucesso.');
+        showToast('PremiacaoReconquista atualizado com sucesso.', 'success');
+
+
 
         const updatedPremiacaoReconquistaes = data.map((item) =>
           item.descricao === editingPremiacaoReconquista.descricao ? { ...item, ...updatedData } : item
@@ -107,45 +125,68 @@ const PremiacaoReconquista = ({ toggleTheme }) => {
         setData(updatedPremiacaoReconquistaes);
         setFilteredData(updatedPremiacaoReconquistaes);
       } catch (error) {
-        setErrorMessage('Erro ao atualizar premiacaoReconquista.');
-        console.error('Erro ao atualizar premiacaoReconquista:', error);
+        showToast('Erro ao atualizar premiacao reconquista.', 'error');
+
+        console.error('Erro ao atualizar Premiação Reconquista:', error);
       } finally {
+        setOpenConfirmDialog(false);
         handleCloseModal();
       }
     }
   };
-
-  const handleDelete = (premiacaoReconquista) => {
-    if (window.confirm('Tem certeza de que deseja excluir este premiacaoReconquista?')) {
-      deletePremiacaoReconquistaById(premiacaoReconquista);
-    }
+  const handleCancelConfirm = () => {
+    setOpenConfirmDialog(false); // Fecha o modal de confirmação
   };
 
-  const deletePremiacaoReconquistaById = async (premiacaoReconquista) => {
+
+  const handleDelete = async () => {
+    if (!itemToDelete) {
+      console.log("Nenhum item para deletar."); // Log se não houver item para deletar
+      return;
+    }
+
+    console.log("Item a ser deletado:", itemToDelete); // Log do item a ser deletado
+
     try {
-      const { descricao, time, valor } = premiacaoReconquista;
+      const { descricao, time, valor } = itemToDelete;
       await deletePremiacaoReconquista({ descricao, time, valor });
-      setData(data.filter((item) => item.id !== premiacaoReconquista.id));
-      setFilteredData(filteredData.filter((item) => item.id !== premiacaoReconquista.id));
-      setSuccessMessage('PremiacaoReconquista deletado com sucesso.');
+
+      const updatedData = data.filter((item) => item.id !== itemToDelete.id);
+      setData(updatedData);
+      setFilteredData(updatedData);
+
+      showToast('Premiação Reconquista deletado com sucesso.', 'success');
     } catch (error) {
-      setErrorMessage('Erro ao excluir premiacaoReconquista.');
-      console.error('Erro ao excluir premiacaoReconquista:', error);
+      showToast('Erro ao excluir Premiação Reconquista.', 'error');
+      console.error('Erro ao excluir Premiação Reconquista:', error);
+    } finally {
+      closeDeleteDialog();
     }
   };
-
 
   const columns = [
-    { field: 'descricao', headerName: 'Descricao', width: 150 },
+    { field: 'descricao', headerName: 'Descricao', width: 180 },
     { field: 'time', headerName: 'Time', width: 150 },
     { field: 'minimo', headerName: 'Minimo', width: 150 },
     { field: 'maximo', headerName: 'Maximo', width: 150 },
     {
-      field: 'valor', headerName: 'Valor', width: 200, valueFormatter: (params) => {
-        const value = params !== undefined ? Number(params) : 0;
-        return `R$ ${value.toFixed(2)}`;
+      field: 'valor',
+      headerName: 'Valor',
+      width: 200,
+      valueFormatter: (params) => {
+        // Convert the value to a number
+        const numberValue = Number(params || 0);
+    
+        // Format with thousands separators and two decimal places
+        const formattedValue = numberValue.toLocaleString('pt-BR', {
+          minimumFractionDigits: 2,
+          maximumFractionDigits: 2
+        });
+    
+        return `R$ ${formattedValue}`;
       }
     },
+    
     {
       field: 'actions',
       headerName: 'Ações',
@@ -208,7 +249,8 @@ const PremiacaoReconquista = ({ toggleTheme }) => {
                 backgroundColor: '#d32f2f',
               }
             }}
-            onClick={() => handleDelete(params.row)}
+            // onClick={() => handleDelete(params.row)}
+            onClick={() => openDeleteDialog(params.row)}
           >
             <FaTrashAlt style={{ fontSize: '1rem', marginRight: 4 }} />
             Excluir
@@ -238,9 +280,9 @@ const PremiacaoReconquista = ({ toggleTheme }) => {
       { header: 'Minimo', dataKey: 'minimo' },
       { header: 'Maximo', dataKey: 'maximo' },
       { header: 'Valor', dataKey: 'valor' },
- 
+
     ];
-    
+
 
     const columnStyles = {
       nome: { cellWidth: 40 }
@@ -248,10 +290,10 @@ const PremiacaoReconquista = ({ toggleTheme }) => {
     const rows = data.map(row => ({
       descricao: row.descricao,
       time: row.time,
-      minimo:row.minimo,
-      maximo:row.maximo,
+      minimo: row.minimo,
+      maximo: row.maximo,
       valor: row.valor,
-     
+
     }));
     const numberFormatter = new Intl.NumberFormat('pt-BR', {
       minimumFractionDigits: 2,
@@ -263,7 +305,7 @@ const PremiacaoReconquista = ({ toggleTheme }) => {
       ...row,
       valor: numberFormatter.format(row.valor), // Format 'valor' field
     }));
-    
+
     autoTable(doc, {
       columns: columns,
       body: formattedRows, // Use formatted rows here
@@ -284,6 +326,16 @@ const PremiacaoReconquista = ({ toggleTheme }) => {
 
     // Salva o PDF
     doc.save('relatorio_premiacao_reconquista.pdf');
+  };
+
+  const openDeleteDialog = (premiacaoReconquista) => {
+    setItemToDelete(premiacaoReconquista); // Armazena o item selecionado
+    setDeleteDialogOpen(true);
+  };
+
+  const closeDeleteDialog = () => {
+    setDeleteDialogOpen(false);
+    setItemToDelete(null); // Reseta o item selecionado ao fechar o diálogo
   };
   return (
     <Box sx={{ display: 'flex' }}>
@@ -335,7 +387,7 @@ const PremiacaoReconquista = ({ toggleTheme }) => {
                 size="small"
               />
             </Grid>
-            
+
 
           </Grid>
         </Paper>
@@ -350,7 +402,7 @@ const PremiacaoReconquista = ({ toggleTheme }) => {
                 backgroundColor: 'darkgreen',
               },
               height: '36px',
-              width: '10%', // Diminuir a largura do botão
+              width: '10%', // Diminuir a largura do botãoF
             }}
           >
             Exportar PDF
@@ -376,8 +428,11 @@ const PremiacaoReconquista = ({ toggleTheme }) => {
         {successMessage && <Alert severity="success" sx={{ mt: 3 }}>{successMessage}</Alert>}
         {errorMessage && <Alert severity="error" sx={{ mt: 3 }}>{errorMessage}</Alert>}
       </Box>
+ 
+
+
       <Dialog open={openModal} onClose={handleCloseModal} fullWidth maxWidth="sm">
-        <DialogTitle>Editar PremiacaoReconquista</DialogTitle>
+        <DialogTitle>Editar Premiação Reconquista</DialogTitle>
         <DialogContent>
           {editingPremiacaoReconquista && (
             <Box sx={{ display: 'flex', flexDirection: 'column', gap: 2, marginTop: 3 }}>
@@ -406,42 +461,114 @@ const PremiacaoReconquista = ({ toggleTheme }) => {
                       borderColor: '#000000',
                     },
                   },
+                  width: '400px',
+                  height: '56px',
+                  borderRadius: '8px',
                 }}
               />
               <TextField
                 label="Time"
                 variant="outlined"
                 value={editingPremiacaoReconquista.time}
-                onChange={(e) => setEditingPremiacaoReconquista({ ...editingPremiacaoReconquista, time: e.target.value })}
+                disabled
                 fullWidth
+                sx={{
+                  '& .MuiInputBase-input': {
+                    color: '#000000',
+                  },
+                  '& .MuiInputLabel-root': {
+                    color: '#000000',
+                  },
+                  '& .MuiOutlinedInput-root': {
+                    backgroundColor: '#f5f5f5',
+                    '& fieldset': {
+                      borderColor: '#000000',
+                    },
+                    '&:hover fieldset': {
+                      borderColor: '#000000',
+                    },
+                    '&.Mui-focused fieldset': {
+                      borderColor: '#000000',
+                    },
+                  },
+                  width: '400px',
+                  height: '56px',
+                  borderRadius: '8px',
+                }}
+              />
+
+
+
+
+
+              <TextField
+                label="Minimo"
+                name="minimo"
+                value={editingPremiacaoReconquista.minimo}
+                onChange={(e) => setEditingPremiacaoReconquista({ ...editingPremiacaoReconquista, minimo: e.target.value })}
+                fullWidth
+                margin="normal"
+                variant="filled"
+                type="number" // Adicionando o tipo number
+                inputProps={{
+                  min: 0, // Opcional: define o valor mínimo como 0
+                  step: 1, // Apenas números inteiros
+                }}
+                sx={{
+                  width: '400px',
+                  height: '56px',
+                  borderRadius: '8px',
+                }}
+              />
+
+              <TextField
+                label="Maximo"
+                name="maximo"
+                value={editingPremiacaoReconquista.maximo}
+                onChange={(e) => setEditingPremiacaoReconquista({ ...editingPremiacaoReconquista, maximo: e.target.value })}
+                fullWidth
+                margin="normal"
+                variant="filled"
+                type="number" // Adicionando o tipo number
+                inputProps={{
+                  min: 0, // Opcional: define o valor mínimo como 0
+                  step: 1, // Apenas números inteiros
+                }}
+                sx={{
+                  width: '400px',
+                  height: '56px',
+                  borderRadius: '8px',
+                }}
               />
 
               <TextField
                 label="Valor"
-                variant="outlined"
+                name="valor"
                 value={editingPremiacaoReconquista.valor}
-                onChange={(e) => setEditingPremiacaoReconquista({ ...editingPremiacaoReconquista, valor: e.target.value })}
-                fullWidth
-              />
-              <TextField
-                label="Minimo"
-                variant="outlined"
-                value={editingPremiacaoReconquista.minimo}
-                onChange={(e) => setEditingPremiacaoReconquista({ ...editingPremiacaoReconquista, minimo: e.target.value })}
-                fullWidth
-              />
-              <TextField
-                label="Maximo"
-                variant="outlined"
-                value={editingPremiacaoReconquista.maximo}
-                onChange={(e) => setEditingPremiacaoReconquista({ ...editingPremiacaoReconquista, maximo: e.target.value })}
-                fullWidth
+                onChange={(e) => {
+                  const value = e.target.value;
+                  // Utiliza expressão regular para permitir apenas números e até duas casas decimais
+                  if (/^\d*\.?\d{0,2}$/.test(value) || value === '') {
+                    setEditingPremiacaoReconquista({ ...editingPremiacaoReconquista, valor: value }); // Atualiza com o valor válido
+                  }
+                }}
+                margin="normal"
+                variant="filled"
+                type="number" // Permite a entrada de números
+                inputProps={{
+                  step: "0.01", // Permite valores decimais
+                }}
+                sx={{
+                  width: '400px',
+                  height: '56px',
+                  borderRadius: '8px',
+                }}
               />
             </Box>
 
           )}
         </DialogContent>
-        <DialogActions>
+        <DialogActions sx={{ display: 'flex', justifyContent: 'space-between', width: '100%' }}>
           <Button
             onClick={handleCloseModal}
             sx={{
@@ -465,6 +592,88 @@ const PremiacaoReconquista = ({ toggleTheme }) => {
             }}
           >
             Salvar
+          </Button>
+        </DialogActions>
+      </Dialog>
+      <Dialog
+        open={deleteDialogOpen}
+        onClose={closeDeleteDialog}
+      >
+        <DialogTitle>Confirmar Delete</DialogTitle>
+        <DialogContent>
+          <DialogContentText>
+            Tem certeza de que deseja deletar "{itemToDelete?.descricao}"?
+          </DialogContentText>
+        </DialogContent>
+        <DialogActions sx={{ display: 'flex', justifyContent: 'space-between', width: '100%' }}>
+          <Box sx={{ flexGrow: 1 }}>
+            <Button
+              onClick={closeDeleteDialog}
+              sx={{
+                backgroundColor: 'red',
+                color: '#fff',
+                '&:hover': {
+                  backgroundColor: '#d32f2f', // Um tom mais escuro de vermelho
+                }
+              }}
+            >
+              Cancelar
+            </Button>
+          </Box>
+          <Box>
+            <Button
+              onClick={handleDelete}
+              sx={{
+                backgroundColor: '#45a049',
+                color: '#fff',
+                '&:hover': {
+                  backgroundColor: '#388e3c', // Um tom mais escuro de verde
+                }
+              }}
+            >
+              Confirmar
+            </Button>
+          </Box>
+        </DialogActions>
+
+      </Dialog>
+
+
+
+      <Dialog
+        open={openConfirmDialog}
+        onClose={handleCancelConfirm}
+        fullWidth
+        maxWidth="xs"
+      >
+        <DialogTitle>Confirmação</DialogTitle>
+        <DialogContent>
+          <p>Tem certeza de que deseja salvar as alterações?</p>
+        </DialogContent>
+        <DialogActions sx={{ display: 'flex', justifyContent: 'space-between', width: '100%' }}>
+          <Button
+            onClick={handleCancelConfirm}
+            sx={{
+              backgroundColor: '#d32f2f',
+              color: '#fff',
+              '&:hover': {
+                backgroundColor: '#d32f2f',
+              }
+            }}
+          >
+            Cancelar
+          </Button>
+          <Button
+            onClick={handleConfirmSave} // Realiza a ação de salvar
+            sx={{
+              backgroundColor: '#45a049',
+              color: '#fff',
+              '&:hover': {
+                backgroundColor: '#388e3c',
+              }
+            }}
+          >
+            Confirmar
           </Button>
         </DialogActions>
       </Dialog>
